Move api file to utilities folder, adjust imports
