+++ src/components/SelectedMovie/SelectedMovie.js
@@ -1,7 +1,7 @@
 // import React from "react"
 import "./SelectedMovie.css"
 import React, { Component} from "react";
-import apiCalls from "../../apiCalls"
+import apiCalls from "../../Utilities/apiCalls"
 import loadingImage from "../../assets/loadingImage.jpg";
 import { Link } from 'react-router-dom'
 import { modifyMovieData } from '../../Utilities/Utils'
